feat(API): revamp package
I swear this is the last time I do this :D

BREAKING CHANGE: The library now exports: import {getTimeZones, rawTimeZones,
timeZonesNames} from "@vvo/tzdb". The explanation for each export is inside the
README.md. tl;dr; you can now either get "raw" time zones with offset and naming
relative to the time zone when not in DST. Or you can get "current" time time
zones information. Which is now a method to call, because it needs to be ran at
runtime.

- added `countryCode`
- renamed `formatted` to `rawFormat`
- added `currentTimeFormat`
- added `currentTimeOffsetInMinutes`