feat(timezones): add deprecated time zones names
This commit adds deprecated time zones names to raw-time-zones.json and the
underlying APIs like getTimeZones.

This does not solves the issue that some deprecated time zones names are used as
main time zones.

Big thanks to @hirusi and @KevinNovak for their help and ideas on how to solve
this.

fixes #44